Use project dictionary for spell checks
1. Split ProjectDict into separate file to avoid circular imports
2. When spell checking, first check if word is an exact match for
a bad word, then check if it (or a variant, such as case) is in the
main dictionary or a good word from the project dictionary.
